12. Numbers have life; they're not just symbols on paper

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -9,9 +9,9 @@
 
 void print_number(int n)
 {
-	long long m; /* power of 10 */
+	int m; /* power of 10 */
 	int c; /* boolean check */
-	long long num; /* convert int to long */
+	int num; /* convert int to long */
 
 	num = n;
 	/* negatives */
